feat(exp): 3.38 and 3.39 grouped to exp.py along with SymExp from other.
